backend/requirements.txt
alembic==1.14.1
celery[redis]==5.4.0
fastapi==0.115.12
httpx==0.28.1
pydantic-settings==2.8.1
psycopg[binary]==3.2.6
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
sqlalchemy==2.0.39
uvicorn[standard]==0.34.0
pytest==8.3.5
pytest-cov==6.0.0

